+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed the SyncCall and AsyncCall traits to use an associated type for their output instead of a generic parameter.
 * Call builders now generally implement `IntoFuture`, allowing `.call_and_wait().await` to be shortened to `.await`.
 
+* Added `log_visibility` to canister creation and canister setting update options.
+
 ## [0.35.0] - 2024-05-10
 
 * Added a limit to the concurrent requests an agent will make at once. This should make server-side ratelimiting much rarer to encounter, even when sending a high volume of requests (for example, a large `ic_utils::ManagementCanister::install` call).